refactor: move logic into its own file
