+++ packages/oas/src/operation/index.ts
@@ -1,11 +1,10 @@
 import type { CallbackExamples } from './lib/get-callback-examples.js';
-import type { MediaTypeExample } from './lib/get-mediatype-examples.js';
 import type { getParametersAsJSONSchemaOptions } from './lib/get-parameters-as-json-schema.js';
 import type { RequestBodyExamples } from './lib/get-requestbody-examples.js';
 import type { ResponseExamples } from './lib/get-response-examples.js';
+import type { Extensions } from '../extensions.js';
 import type { OpenAPIV3, OpenAPIV3_1 } from 'openapi-types';
 
-import { getExtension, type Extensions } from '../extensions.js';
 import findSchemaDefinition from '../lib/find-schema-definition.js';
 import matchesMimeType from '../lib/matches-mimetype.js';
 import * as RMOAS from '../types.js';
@@ -13,6 +12,7 @@ import { supportedMethods } from '../utils.js';
 
 import { dedupeCommonParameters } from './lib/dedupe-common-parameters.js';
 import { getCallbackExamples } from './lib/get-callback-examples.js';
+import { getExampleGroups, type ExampleGroups } from './lib/get-example-groups.js';
 import { getParametersAsJSONSchema } from './lib/get-parameters-as-json-schema.js';
 import { getRequestBodyExamples } from './lib/get-requestbody-examples.js';
 import { getResponseAsJSONSchema } from './lib/get-response-as-json-schema.js';
@@ -800,182 +800,19 @@ export class Operation {
     return this.schema?.[extension];
   }
 
+  /**
+   * TKTK
+   */
   getExampleGroups(): ExampleGroups {
     if (this.exampleGroups) return this.exampleGroups;
 
-    const pairs: ExampleGroups = {};
-
-    // first, parse through custom code samples
-    const codeSamples = getExtension('code-samples', this.api, this) as Extensions['code-samples'][];
-
-    let namelessCodeSamples = 0;
-    // internal key to represent code samples that do not have a corresponding response example,
-    // but should still be surfaced in the UI
-    const noCorrespondingResponseKey = 'NoCorrespondingResponseForCustomCodeSample';
-    // add any and all code samples since they take precedence
-    (codeSamples || [])?.forEach(sample => {
-      namelessCodeSamples += 1;
-      // sample contains `correspondingExample` key and key already exists in pair object
-      if (pairs[sample.correspondingExample]?.customCodeSamples?.length) {
-        // append code sample to existing list of items
-        pairs[sample.correspondingExample].customCodeSamples.push(sample);
-      }
-      // sample contains `correspondingExample` key (fallback)
-      else if (sample.correspondingExample) {
-        const name =
-          sample.name && sample.name.length > 0
-            ? sample.name
-            : `Default${namelessCodeSamples > 1 ? ` #${namelessCodeSamples}` : ''}`;
-
-        // create example pair entry with code sample
-        pairs[sample.correspondingExample] = {
-          name,
-          customCodeSamples: [sample],
-        };
-      }
-      // sample does not contain corresponding response example and internal key already exists
-      else if (pairs[noCorrespondingResponseKey]?.customCodeSamples?.length) {
-        // append code sample to existing list of items
-        pairs[noCorrespondingResponseKey].customCodeSamples.push(sample);
-      }
-      // sample does not contain corresponding response example (fallback)
-      else {
-        const name =
-          sample.name && sample.name.length > 0
-            ? sample.name
-            : `Default${namelessCodeSamples > 1 ? ` #${namelessCodeSamples}` : ''}`;
-
-        // create example pair entry with code sample
-        pairs[noCorrespondingResponseKey] = {
-          name,
-          customCodeSamples: [sample],
-        };
-      }
-    });
-
-    // if code samples with `correspondingExample` delineator exist, add those
-    // and add any matching responses and then return
-    if (Object.keys(pairs).length) {
-      this.getResponseExamples().forEach(example => {
-        Object.entries(example.mediaTypes || {}).forEach(([mediaType, mediaTypeExamples]) => {
-          mediaTypeExamples.forEach(mediaTypeExample => {
-            if (mediaTypeExample.title && Object.keys(pairs).includes(mediaTypeExample.title)) {
-              pairs[mediaTypeExample.title].response = {
-                mediaType,
-                mediaTypeExample,
-                status: example.status,
-              };
-
-              // if the current pair doesn't already have a name set, use the response example summary
-              if (!pairs[mediaTypeExample.title].name) {
-                pairs[mediaTypeExample.title].name = mediaTypeExample.summary;
-              }
-            }
-          });
-        });
-      });
-    } else {
-      // if no custom code examples, parse through param and body examples
-      this.getParameters().forEach(param => {
-        Object.entries(param.examples || {}).forEach(
-          ([exampleKey, paramExample]: [string, OpenAPIV3.ExampleObject]) => {
-            pairs[exampleKey] = {
-              ...pairs[exampleKey],
-              name: paramExample.summary,
-              request: {
-                ...pairs[exampleKey]?.request,
-                [param.in]: {
-                  ...pairs[exampleKey]?.request?.[param.in],
-                  [param.name]: paramExample.value,
-                },
-              },
-            };
-          },
-        );
-      });
-
-      // add request body examples
-      this.getRequestBodyExamples().forEach(requestExample => {
-        requestExample.examples.forEach((mediaTypeExample: MediaTypeExample) => {
-          if (mediaTypeExample.title && Object.keys(pairs).includes(mediaTypeExample.title)) {
-            pairs[mediaTypeExample.title] = {
-              ...pairs[mediaTypeExample.title],
-              request: {
-                ...pairs[mediaTypeExample.title]?.request,
-                body: mediaTypeExample.value,
-              },
-            };
-
-            // if the current pair doesn't already have a name set, use the response example summary
-            if (!pairs[mediaTypeExample.title].name) {
-              pairs[mediaTypeExample.title].name = mediaTypeExample.summary;
-            }
-          }
-        });
-      });
-
-      if (Object.keys(pairs).length) {
-        // if there are matching keys, add the corresponding response examples
-        this.getResponseExamples().forEach(example => {
-          Object.entries(example.mediaTypes || {}).forEach(([mediaType, mediaTypeExamples]) => {
-            mediaTypeExamples.forEach(mediaTypeExample => {
-              if (mediaTypeExample.title && Object.keys(pairs).includes(mediaTypeExample.title)) {
-                pairs[mediaTypeExample.title].response = {
-                  mediaType,
-                  mediaTypeExample,
-                  status: example.status,
-                };
-
-                // if the current pair doesn't already have a name set, use the response example summary
-                if (!pairs[mediaTypeExample.title].name) {
-                  pairs[mediaTypeExample.title].name = mediaTypeExample.summary;
-                }
-              }
-            });
-          });
-        });
-      }
-    }
-
-    // cull any objects that don't have both request + response
-    Object.entries(pairs).forEach(([pairId, pair]) => {
-      if (pair.request && !pair.response) {
-        delete pairs[pairId];
-      }
-    });
+    const pairs = getExampleGroups(this);
 
     this.exampleGroups = pairs;
 
     return pairs;
   }
 }
-
-type ExampleGroups = Record<
-  string,
-  {
-    /**
-     * List of custom code samples that contain `correspondingExample` key.
-     * Mutually exclusive of `request`.
-     */
-    customCodeSamples?: Extensions['code-samples'][];
-    /**
-     * Title of example pair. Precedence is as follows: TKTK
-     */
-    name: string;
-    /**
-     * TKTK
-     */
-    request?: RMOAS.DataForHAR;
-    /**
-     * TKTK
-     */
-    response?: {
-      mediaType: string;
-      mediaTypeExample: MediaTypeExample;
-      status: string;
-    };
-  }
->;
 
 export class Callback extends Operation {
   /**
